+++ packages/blade/src/components/Dropdown/Dropdown.tsx
@@ -28,6 +28,7 @@ const validDropdownChildren = [
   dropdownComponentIds.DropdownOverlay,
   dropdownComponentIds.triggers.AutoComplete,
   bottomSheetComponentIds.BottomSheet,
+  dropdownComponentIds.triggers.FilterChipSelectInput,
 ];
 
 /**
@@ -80,10 +81,7 @@ const _Dropdown = (
   const [activeTagIndex, setActiveTagIndex] = React.useState(-1);
   const [shouldIgnoreBlurAnimation, setShouldIgnoreBlurAnimation] = React.useState(false);
   const [hasFooterAction, setHasFooterAction] = React.useState(false);
-  const [
-    hasAutoCompleteInBottomSheetHeader,
-    setHasAutoCompleteInBottomSheetHeader,
-  ] = React.useState(false);
+  const [hasAutoCompleteInHeader, setHasAutoCompleteInHeader] = React.useState(false);
   const [isKeydownPressed, setIsKeydownPressed] = React.useState(false);
   const [changeCallbackTriggerer, setChangeCallbackTriggerer] = React.useState<
     DropdownContextType['changeCallbackTriggerer']
@@ -98,6 +96,7 @@ const _Dropdown = (
    * */
   const triggererWrapperRef = React.useRef<ContainerElementType>(null);
   const triggererRef = React.useRef<HTMLButtonElement>(null);
+  const headerAutoCompleteRef = React.useRef<HTMLButtonElement>(null);
   const actionListItemRef = React.useRef<HTMLDivElement>(null);
   const dropdownTriggerer = React.useRef<DropdownContextType['dropdownTriggerer']>();
   const isTagDismissedRef = React.useRef<{ value: boolean } | null>({ value: false });
@@ -161,6 +160,9 @@ const _Dropdown = (
       if (isValidAllowedChildren(child, dropdownComponentIds.triggers.AutoComplete)) {
         dropdownTriggerer.current = 'AutoComplete';
       }
+      if (isValidAllowedChildren(child, dropdownComponentIds.triggers.FilterChipSelectInput)) {
+        dropdownTriggerer.current = 'FilterChipSelectInput';
+      }
     }
   });
 
@@ -188,13 +190,14 @@ const _Dropdown = (
       setIsKeydownPressed,
       dropdownBaseId,
       triggererRef,
+      headerAutoCompleteRef,
       triggererWrapperRef,
       actionListItemRef,
       selectionType,
       hasFooterAction,
       setHasFooterAction,
-      hasAutoCompleteInBottomSheetHeader,
-      setHasAutoCompleteInBottomSheetHeader,
+      hasAutoCompleteInHeader,
+      setHasAutoCompleteInHeader,
       dropdownTriggerer: dropdownTriggerer.current,
       changeCallbackTriggerer,
       setChangeCallbackTriggerer,
@@ -225,13 +228,13 @@ const _Dropdown = (
     return {
       isOpen: isDropdownOpen,
       dropdownHasBottomSheet,
-      hasAutoCompleteInBottomSheetHeader,
+      hasAutoCompleteInHeader,
       setDropdownHasBottomSheet,
       // This is the dismiss function which will be injected into the BottomSheet
       // Basically <BottomSheet onDismiss={onBottomSheetDismiss} />
       onBottomSheetDismiss: close,
     };
-  }, [dropdownHasBottomSheet, hasAutoCompleteInBottomSheetHeader, isDropdownOpen, close]);
+  }, [dropdownHasBottomSheet, hasAutoCompleteInHeader, isDropdownOpen, close]);
 
   return (
     <BottomSheetAndDropdownGlueContext.Provider value={BottomSheetAndDropdownGlueContextValue}>