+++ packages/blade/src/components/Dropdown/index.ts
@@ -5,3 +5,5 @@ export * from './DropdownLink';
 export * from './DropdownIconButton';
 export * from './DropdownHeaderFooter';
 export * from './types';
+export * from './FilterChipSelectInput';
+export * from './FilterChipGroup';