+++ packages/blade/src/components/QuickFilters/QuickFilterGroup/index.tsx
@@ -1,0 +1,2 @@
+export { QuickFilterGroup } from './QuickFilterGroup';
+export { useQuickFilterGroupContext, QuickFilterGroupProvider } from './QuickFilterContext';